Refactoring for processor .h / .cpp
