+++ src/processor/incremental_triangulator.cpp
@@ -1,0 +1,666 @@
+#include "incremental_triangulator.h"
+
+void IncrementalTriangulator::Options::Check() const {
+}
+
+IncrementalTriangulator::IncrementalTriangulator(const SceneGraph* scene_graph, Reconstruction* reconstruction)
+        : scene_graph_(scene_graph), reconstruction_(reconstruction) { }
+
+size_t IncrementalTriangulator::TriangulateImage(const Options& options, const image_t image_id) {
+    options.Check();
+
+    size_t num_tris = 0;
+
+    ClearCaches();
+
+    const Image& image = reconstruction_->Image(image_id);
+    if (!image.IsRegistered()) {
+        return num_tris;
+    }
+
+    const Camera& camera = reconstruction_->Camera(image.CameraId());
+    if (HasCameraBogusParams(options, camera)) {
+        return num_tris;
+    }
+
+    CorrData ref_corr_data;
+    ref_corr_data.image_id = image_id;
+    ref_corr_data.image = &image;
+    ref_corr_data.camera = &camera;
+    ref_corr_data.proj_matrix = image.ProjectionMatrix();
+
+    std::vector<CorrData> corrs_data;
+
+    for (point2D_t point2D_idx = 0; point2D_idx < image.NumPoints2D();
+         ++point2D_idx) {
+        const size_t num_triangulated =
+                Find(options, image_id, point2D_idx,
+                     static_cast<size_t>(options.max_transitivity), &corrs_data);
+        if (corrs_data.empty()) {
+            continue;
+        }
+
+        const Point2D& point2D = image.Point2D(point2D_idx);
+        ref_corr_data.point2D_idx = point2D_idx;
+        ref_corr_data.point2D = &point2D;
+
+        if (num_triangulated == 0) {
+            corrs_data.push_back(ref_corr_data);
+            num_tris += Create(options, corrs_data);
+        } else {
+            num_tris += Continue(options, ref_corr_data, corrs_data);
+            corrs_data.push_back(ref_corr_data);
+            num_tris += Create(options, corrs_data);
+        }
+    }
+
+    return num_tris;
+}
+
+size_t IncrementalTriangulator::CompleteImage(const Options& options, const image_t image_id) {
+    options.Check();
+
+    size_t num_tris = 0;
+
+    ClearCaches();
+
+    const Image& image = reconstruction_->Image(image_id);
+    if (!image.IsRegistered()) {
+        return num_tris;
+    }
+
+    const Camera& camera = reconstruction_->Camera(image.CameraId());
+    if (HasCameraBogusParams(options, camera)) {
+        return num_tris;
+    }
+
+    CorrData ref_corr_data;
+    ref_corr_data.image_id = image_id;
+    ref_corr_data.image = &image;
+    ref_corr_data.camera = &camera;
+    ref_corr_data.proj_matrix = image.ProjectionMatrix();
+
+    std::vector<CorrData> corrs_data;
+
+    for (point2D_t point2D_idx = 0; point2D_idx < image.NumPoints2D();
+         ++point2D_idx) {
+        const Point2D& point2D = image.Point2D(point2D_idx);
+        if (point2D.HasPoint3D()) {
+            num_tris += Complete(options, point2D.Point3DId());
+            continue;
+        }
+
+        if (options.ignore_two_view_tracks &&
+            scene_graph_->IsTwoViewObservation(image_id, point2D_idx)) {
+            continue;
+        }
+
+        const size_t num_triangulated =
+                Find(options, image_id, point2D_idx,
+                     static_cast<size_t>(options.max_transitivity), &corrs_data);
+        if (num_triangulated || corrs_data.empty()) {
+            continue;
+        }
+
+        ref_corr_data.point2D = &point2D;
+        ref_corr_data.point2D_idx = point2D_idx;
+        corrs_data.push_back(ref_corr_data);
+
+        std::vector<TriangulationEstimator::PointData> point_data;
+        point_data.resize(corrs_data.size());
+        std::vector<TriangulationEstimator::PoseData> pose_data;
+        pose_data.resize(corrs_data.size());
+        for (size_t i = 0; i < corrs_data.size(); ++i) {
+            const CorrData& corr_data = corrs_data[i];
+            point_data[i].point = corr_data.point2D->XY();
+            point_data[i].point_normalized =
+                    corr_data.camera->ImageToWorld(point_data[i].point);
+            pose_data[i].proj_matrix = corr_data.image->ProjectionMatrix();
+            pose_data[i].proj_center = corr_data.image->ProjectionCenter();
+            pose_data[i].camera = corr_data.camera;
+        }
+
+        EstimateTriangulationOptions tri_options;
+        tri_options.min_tri_angle = DegToRad(options.min_angle);
+        tri_options.residual_type =
+                TriangulationEstimator::ResidualType::REPROJECTION_ERROR;
+        tri_options.ransac_options.max_error = options.complete_max_reproj_error;
+        tri_options.ransac_options.confidence = 0.9999;
+        tri_options.ransac_options.min_inlier_ratio = 0.02;
+        tri_options.ransac_options.max_num_trials = 10000;
+
+        const size_t kExhaustiveSamplingThreshold = 15;
+        if (point_data.size() <= kExhaustiveSamplingThreshold) {
+            tri_options.ransac_options.min_num_trials =
+                    NChooseK(point_data.size(), 2);
+        }
+
+        Eigen::Vector3d xyz;
+        std::vector<bool> inlier_mask;
+        if (!EstimateTriangulation(tri_options, point_data, pose_data, &inlier_mask,
+                                   &xyz)) {
+            continue;
+        }
+
+        Track track;
+        track.Reserve(corrs_data.size());
+        for (size_t i = 0; i < inlier_mask.size(); ++i) {
+            if (inlier_mask[i]) {
+                const CorrData& corr_data = corrs_data[i];
+                track.AddElement(corr_data.image_id, corr_data.point2D_idx);
+                num_tris += 1;
+            }
+        }
+
+        const point3D_t point3D_id = reconstruction_->AddPoint3D(xyz, track);
+        changed_point3D_ids_.insert(point3D_id);
+    }
+
+    return num_tris;
+}
+
+size_t IncrementalTriangulator::CompleteTracks(const Options& options,
+                                               const std::unordered_set<point3D_t>& point3D_ids) {
+    options.Check();
+
+    size_t num_completed = 0;
+
+    ClearCaches();
+
+    for (const point3D_t point3D_id : point3D_ids) {
+        num_completed += Complete(options, point3D_id);
+    }
+
+    return num_completed;
+}
+
+size_t IncrementalTriangulator::CompleteAllTracks(const Options& options) {
+    options.Check();
+
+    size_t num_completed = 0;
+
+    ClearCaches();
+
+    for (const point3D_t point3D_id : reconstruction_->Point3DIds()) {
+        num_completed += Complete(options, point3D_id);
+    }
+
+    return num_completed;
+}
+
+size_t IncrementalTriangulator::MergeTracks(const Options& options,
+                                            const std::unordered_set<point3D_t>& point3D_ids) {
+    options.Check();
+
+    size_t num_merged = 0;
+
+    ClearCaches();
+
+    for (const point3D_t point3D_id : point3D_ids) {
+        num_merged += Merge(options, point3D_id);
+    }
+
+    return num_merged;
+}
+
+size_t IncrementalTriangulator::MergeAllTracks(const Options& options) {
+    options.Check();
+
+    size_t num_merged = 0;
+
+    ClearCaches();
+
+    for (const point3D_t point3D_id : reconstruction_->Point3DIds()) {
+        num_merged += Merge(options, point3D_id);
+    }
+
+    return num_merged;
+}
+
+size_t IncrementalTriangulator::Retriangulate(const Options& options) {
+    options.Check();
+
+    size_t num_tris = 0;
+
+    ClearCaches();
+
+    Options re_options = options;
+    re_options.continue_max_angle_error = options.re_max_angle_error;
+
+    for (const auto& image_pair : reconstruction_->ImagePairs()) {
+        const double tri_ratio =
+                static_cast<double>(image_pair.second.first) / image_pair.second.second;
+        if (tri_ratio >= options.re_min_ratio) {
+            continue;
+        }
+
+        image_t image_id1;
+        image_t image_id2;
+        Database::PairIdToImagePair(image_pair.first, &image_id1, &image_id2);
+
+        const Image& image1 = reconstruction_->Image(image_id1);
+        if (!image1.IsRegistered()) {
+            continue;
+        }
+
+        const Image& image2 = reconstruction_->Image(image_id2);
+        if (!image2.IsRegistered()) {
+            continue;
+        }
+
+        int& num_re_trials = re_num_trials_[image_pair.first];
+        if (num_re_trials >= options.re_max_trials) {
+            continue;
+        }
+        num_re_trials += 1;
+
+        const Eigen::Matrix3x4d proj_matrix1 = image1.ProjectionMatrix();
+        const Eigen::Matrix3x4d proj_matrix2 = image2.ProjectionMatrix();
+
+        const Camera& camera1 = reconstruction_->Camera(image1.CameraId());
+        const Camera& camera2 = reconstruction_->Camera(image2.CameraId());
+        if (HasCameraBogusParams(options, camera1) ||
+            HasCameraBogusParams(options, camera2)) {
+            continue;
+        }
+
+        const std::vector<std::pair<point2D_t, point2D_t>> corrs =
+                scene_graph_->FindCorrespondencesBetweenImages(image_id1, image_id2);
+
+        for (const auto& corr : corrs) {
+            const Point2D& point2D1 = image1.Point2D(corr.first);
+            const Point2D& point2D2 = image2.Point2D(corr.second);
+
+            if (point2D1.HasPoint3D() && point2D2.HasPoint3D()) {
+                continue;
+            }
+
+            CorrData corr_data1;
+            corr_data1.image_id = image_id1;
+            corr_data1.point2D_idx = corr.first;
+            corr_data1.image = &image1;
+            corr_data1.camera = &camera1;
+            corr_data1.point2D = &point2D1;
+            corr_data1.proj_matrix = proj_matrix1;
+
+            CorrData corr_data2;
+            corr_data2.image_id = image_id2;
+            corr_data2.point2D_idx = corr.second;
+            corr_data2.image = &image2;
+            corr_data2.camera = &camera2;
+            corr_data2.point2D = &point2D2;
+            corr_data2.proj_matrix = proj_matrix2;
+
+            if (point2D1.HasPoint3D() && !point2D2.HasPoint3D()) {
+                const std::vector<CorrData> corrs_data1 = {corr_data1};
+                num_tris += Continue(re_options, corr_data2, corrs_data1);
+            } else if (!point2D1.HasPoint3D() && point2D2.HasPoint3D()) {
+                const std::vector<CorrData> corrs_data2 = {corr_data2};
+                num_tris += Continue(re_options, corr_data1, corrs_data2);
+            } else if (!point2D1.HasPoint3D() && !point2D2.HasPoint3D()) {
+                const std::vector<CorrData> corrs_data = {corr_data1, corr_data2};
+                num_tris += Create(options, corrs_data);
+            }
+        }
+    }
+
+    return num_tris;
+}
+
+std::unordered_set<point3D_t> IncrementalTriangulator::ChangedPoints3D() const {
+    std::unordered_set<point3D_t> point3D_ids;
+    point3D_ids.reserve(changed_point3D_ids_.size());
+
+    for (const point3D_t point3D_id : changed_point3D_ids_) {
+        if (reconstruction_->ExistsPoint3D(point3D_id)) {
+            point3D_ids.insert(point3D_id);
+        }
+    }
+
+    return point3D_ids;
+}
+
+void IncrementalTriangulator::ClearChangedPoints3D() {
+    changed_point3D_ids_.clear();
+}
+
+void IncrementalTriangulator::ClearCaches() {
+    camera_has_bogus_params_.clear();
+    merge_trials_.clear();
+}
+
+size_t IncrementalTriangulator::Find(const Options& options,
+                                     const image_t image_id,
+                                     const point2D_t point2D_idx,
+                                     const size_t transitivity,
+                                     std::vector<CorrData>* corrs_data) {
+    const std::vector<SceneGraph::Correspondence>& corrs =
+            scene_graph_->FindTransitiveCorrespondences(image_id, point2D_idx,
+                                                        transitivity);
+
+    corrs_data->clear();
+    corrs_data->reserve(corrs.size());
+
+    size_t num_triangulated = 0;
+
+    for (const SceneGraph::Correspondence corr : corrs) {
+        const Image& corr_image = reconstruction_->Image(corr.image_id);
+        if (!corr_image.IsRegistered()) {
+            continue;
+        }
+
+        const Camera& corr_camera = reconstruction_->Camera(corr_image.CameraId());
+        if (HasCameraBogusParams(options, corr_camera)) {
+            continue;
+        }
+
+        CorrData corr_data;
+        corr_data.image_id = corr.image_id;
+        corr_data.point2D_idx = corr.point2D_idx;
+        corr_data.image = &corr_image;
+        corr_data.camera = &corr_camera;
+        corr_data.point2D = &corr_image.Point2D(corr.point2D_idx);
+        corr_data.proj_matrix = corr_image.ProjectionMatrix();
+
+        corrs_data->push_back(corr_data);
+
+        if (corr_data.point2D->HasPoint3D()) {
+            num_triangulated += 1;
+        }
+    }
+
+    return num_triangulated;
+}
+
+size_t IncrementalTriangulator::Create(const Options& options,
+                                       const std::vector<CorrData>& corrs_data) {
+    std::vector<CorrData> create_corrs_data;
+    create_corrs_data.reserve(corrs_data.size());
+    for (const CorrData& corr_data : corrs_data) {
+        if (!corr_data.point2D->HasPoint3D()) {
+            create_corrs_data.push_back(corr_data);
+        }
+    }
+
+    if (create_corrs_data.size() < 2) {
+        return 0;
+    } else if (options.ignore_two_view_tracks && create_corrs_data.size() == 2) {
+        const CorrData& corr_data1 = create_corrs_data[0];
+        if (scene_graph_->IsTwoViewObservation(corr_data1.image_id,
+                                               corr_data1.point2D_idx)) {
+            return 0;
+        }
+    }
+
+    std::vector<TriangulationEstimator::PointData> point_data;
+    point_data.resize(create_corrs_data.size());
+    std::vector<TriangulationEstimator::PoseData> pose_data;
+    pose_data.resize(create_corrs_data.size());
+    for (size_t i = 0; i < create_corrs_data.size(); ++i) {
+        const CorrData& corr_data = create_corrs_data[i];
+        point_data[i].point = corr_data.point2D->XY();
+        point_data[i].point_normalized =
+                corr_data.camera->ImageToWorld(point_data[i].point);
+        pose_data[i].proj_matrix = corr_data.image->ProjectionMatrix();
+        pose_data[i].proj_center = corr_data.image->ProjectionCenter();
+        pose_data[i].camera = corr_data.camera;
+    }
+
+    EstimateTriangulationOptions tri_options;
+    tri_options.min_tri_angle = DegToRad(options.min_angle);
+    tri_options.residual_type =
+            TriangulationEstimator::ResidualType::ANGULAR_ERROR;
+    tri_options.ransac_options.max_error =
+            DegToRad(options.create_max_angle_error);
+    tri_options.ransac_options.confidence = 0.9999;
+    tri_options.ransac_options.min_inlier_ratio = 0.02;
+    tri_options.ransac_options.max_num_trials = 10000;
+
+    const size_t kExhaustiveSamplingThreshold = 15;
+    if (point_data.size() <= kExhaustiveSamplingThreshold) {
+        tri_options.ransac_options.min_num_trials = NChooseK(point_data.size(), 2);
+    }
+
+    Eigen::Vector3d xyz;
+    std::vector<bool> inlier_mask;
+    if (!EstimateTriangulation(tri_options, point_data, pose_data, &inlier_mask,
+                               &xyz)) {
+        return 0;
+    }
+
+    Track track;
+    track.Reserve(create_corrs_data.size());
+    for (size_t i = 0; i < inlier_mask.size(); ++i) {
+        if (inlier_mask[i]) {
+            const CorrData& corr_data = create_corrs_data[i];
+            track.AddElement(corr_data.image_id, corr_data.point2D_idx);
+        }
+    }
+
+    const point3D_t point3D_id = reconstruction_->AddPoint3D(xyz, track);
+    changed_point3D_ids_.insert(point3D_id);
+
+    const size_t kMinRecursiveTrackLength = 3;
+    if (create_corrs_data.size() - track.Length() >= kMinRecursiveTrackLength) {
+        return track.Length() + Create(options, create_corrs_data);
+    }
+
+    return track.Length();
+}
+
+size_t IncrementalTriangulator::Continue(const Options& options,
+                                         const CorrData& ref_corr_data,
+                                         const std::vector<CorrData>& corrs_data) {
+    if (ref_corr_data.point2D->HasPoint3D()) {
+        return 0;
+    }
+
+    double best_angle_error = std::numeric_limits<double>::max();
+    size_t best_idx = std::numeric_limits<size_t>::max();
+
+    for (size_t idx = 0; idx < corrs_data.size(); ++idx) {
+        const CorrData& corr_data = corrs_data[idx];
+        if (!corr_data.point2D->HasPoint3D()) {
+            continue;
+        }
+
+        const Point_3D& point3D =
+                reconstruction_->Point3D(corr_data.point2D->Point3DId());
+
+        if (!HasPointPositiveDepth(ref_corr_data.proj_matrix, point3D.XYZ())) {
+            continue;
+        }
+
+        const double angle_error =
+                CalculateAngularError(ref_corr_data.point2D->XY(), point3D.XYZ(),
+                                      ref_corr_data.proj_matrix, *ref_corr_data.camera);
+        if (angle_error < best_angle_error) {
+            best_angle_error = angle_error;
+            best_idx = idx;
+        }
+    }
+
+    const double max_angle_error = DegToRad(options.continue_max_angle_error);
+    if (best_angle_error <= max_angle_error &&
+        best_idx != std::numeric_limits<size_t>::max()) {
+        const CorrData& corr_data = corrs_data[best_idx];
+        const TrackElement track_el(ref_corr_data.image_id,
+                                    ref_corr_data.point2D_idx);
+        reconstruction_->AddObservation(corr_data.point2D->Point3DId(), track_el);
+        changed_point3D_ids_.insert(corr_data.point2D->Point3DId());
+        return 1;
+    }
+
+    return 0;
+}
+
+size_t IncrementalTriangulator::Merge(const Options& options,
+                                      const point3D_t point3D_id) {
+    size_t num_merged = 0;
+
+    if (!reconstruction_->ExistsPoint3D(point3D_id)) {
+        return num_merged;
+    }
+
+    const auto& point3D = reconstruction_->Point3D(point3D_id);
+
+    for (const auto& track_el : point3D.Track().Elements()) {
+        const std::vector<SceneGraph::Correspondence>& corrs =
+                scene_graph_->FindCorrespondences(track_el.image_id,
+                                                  track_el.point2D_idx);
+
+        for (const auto corr : corrs) {
+            Image& image = reconstruction_->Image(corr.image_id);
+            if (!image.IsRegistered()) {
+                continue;
+            }
+
+            const Point2D& corr_point2D = image.Point2D(corr.point2D_idx);
+            if (!corr_point2D.HasPoint3D() ||
+                corr_point2D.Point3DId() == point3D_id ||
+                merge_trials_[point3D_id].count(corr_point2D.Point3DId()) > 0) {
+                continue;
+            }
+
+            const Point_3D& corr_point3D =
+                    reconstruction_->Point3D(corr_point2D.Point3DId());
+
+            merge_trials_[point3D_id].insert(corr_point2D.Point3DId());
+            merge_trials_[corr_point2D.Point3DId()].insert(point3D_id);
+
+            const Eigen::Vector3d merged_xyz =
+                    (point3D.Track().Length() * point3D.XYZ() +
+                     corr_point3D.Track().Length() * corr_point3D.XYZ()) /
+                    (point3D.Track().Length() + corr_point3D.Track().Length());
+
+            size_t num_inliers = 0;
+            for (const Track* track : {&point3D.Track(), &corr_point3D.Track()}) {
+                for (const auto test_track_el : track->Elements()) {
+                    const Image& test_image =
+                            reconstruction_->Image(test_track_el.image_id);
+                    const Camera& test_camera =
+                            reconstruction_->Camera(test_image.CameraId());
+                    const Point2D& test_point2D =
+                            test_image.Point2D(test_track_el.point2D_idx);
+
+                    const Eigen::Matrix3x4d test_proj_matrix =
+                            test_image.ProjectionMatrix();
+
+                    if (HasPointPositiveDepth(test_proj_matrix, merged_xyz) &&
+                        CalculateReprojectionError(test_point2D.XY(), merged_xyz,
+                                                   test_proj_matrix, test_camera) <=
+                        options.merge_max_reproj_error) {
+                        num_inliers += 1;
+                    } else {
+                        break;
+                    }
+                }
+            }
+
+            if (num_inliers ==
+                point3D.Track().Length() + corr_point3D.Track().Length()) {
+                num_merged += num_inliers;
+
+                const point3D_t merged_point3D_id = reconstruction_->MergePoints3D(
+                        point3D_id, corr_point2D.Point3DId());
+
+                changed_point3D_ids_.erase(point3D_id);
+                changed_point3D_ids_.erase(corr_point2D.Point3DId());
+                changed_point3D_ids_.insert(merged_point3D_id);
+
+                return num_merged + Merge(options, merged_point3D_id);
+            }
+        }
+    }
+
+    return num_merged;
+}
+
+size_t IncrementalTriangulator::Complete(const Options& options,
+                                         const point3D_t point3D_id) {
+    size_t num_completed = 0;
+
+    if (!reconstruction_->ExistsPoint3D(point3D_id)) {
+        return num_completed;
+    }
+
+    const Point_3D& point3D = reconstruction_->Point3D(point3D_id);
+
+    std::vector<TrackElement> queue;
+    queue.reserve(point3D.Track().Length());
+
+    for (const auto& track_el : point3D.Track().Elements()) {
+        queue.emplace_back(track_el.image_id, track_el.point2D_idx);
+    }
+
+    const int max_transitivity = options.complete_max_transitivity;
+    for (int transitivity = 0; transitivity < max_transitivity; ++transitivity) {
+        if (queue.empty()) {
+            break;
+        }
+
+        const auto prev_queue = queue;
+        queue.clear();
+
+        for (const TrackElement queue_elem : prev_queue) {
+            const std::vector<SceneGraph::Correspondence>& corrs =
+                    scene_graph_->FindCorrespondences(queue_elem.image_id,
+                                                      queue_elem.point2D_idx);
+
+            for (const auto corr : corrs) {
+                if (corr.image_id == queue_elem.image_id) {
+                    continue;
+                }
+
+                const Image& image = reconstruction_->Image(corr.image_id);
+                if (!image.IsRegistered()) {
+                    continue;
+                }
+
+                const Point2D& point2D = image.Point2D(corr.point2D_idx);
+                if (point2D.HasPoint3D()) {
+                    continue;
+                }
+
+                const Eigen::Matrix3x4d proj_matrix = image.ProjectionMatrix();
+                if (!HasPointPositiveDepth(proj_matrix, point3D.XYZ())) {
+                    continue;
+                }
+
+                const Camera& camera = reconstruction_->Camera(image.CameraId());
+                if (HasCameraBogusParams(options, camera)) {
+                    continue;
+                }
+
+                const double reproj_error = CalculateReprojectionError(
+                        point2D.XY(), point3D.XYZ(), proj_matrix, camera);
+                if (reproj_error > options.complete_max_reproj_error) {
+                    continue;
+                }
+
+                const TrackElement track_el(corr.image_id, corr.point2D_idx);
+                reconstruction_->AddObservation(point3D_id, track_el);
+                changed_point3D_ids_.insert(point3D_id);
+
+                queue.emplace_back(corr.image_id, corr.point2D_idx);
+
+                num_completed += 1;
+            }
+        }
+    }
+
+    return num_completed;
+}
+
+bool IncrementalTriangulator::HasCameraBogusParams(const Options& options,
+                                                   const Camera& camera) {
+    if (camera_has_bogus_params_.count(camera.CameraId()) == 0) {
+        const bool has_bogus_params = camera.HasBogusParams(
+                options.min_focal_length_ratio, options.max_focal_length_ratio,
+                options.max_extra_param);
+        camera_has_bogus_params_.emplace(camera.CameraId(), has_bogus_params);
+        return has_bogus_params;
+    } else {
+        return camera_has_bogus_params_.at(camera.CameraId());
+    }
+}
